Split classes in separated files.
