+++ PokeApi.NET/PokeApi.NET.csproj
@@ -57,6 +57,7 @@
     <Compile Include="API types\Ability.cs" />
     <Compile Include="API types\Description.cs" />
     <Compile Include="API types\EvYield.cs" />
+    <Compile Include="API types\LearnableMove.cs" />
     <Compile Include="API types\Pokedex.cs" />
     <Compile Include="DataFetcher.cs" />
     <Compile Include="Enums.cs" />
